+++ packages/swapkit/helpers/src/modules/__tests__/swapKitNumber.test.ts
@@ -230,6 +230,9 @@ describe('SwapKitNumber', () => {
           currencyPosition: 'end',
         }),
       ).toBe('0,000057€');
+
+      const skNumber4 = new SwapKitNumber(12345);
+      expect(skNumber4.toCurrency()).toBe('$12,345');
     });
   });
 
@@ -325,7 +328,6 @@ describe('SwapKitNumber', () => {
       const result = skNumber1.mul(skNumber2);
 
       // The exact result of 1.23 * 4.56 is 5.6088
-      // If we round it to 2 decimal places, we should get 5.61
       expect(result.getValue('string')).toBe('5.609');
       expect(result.getBaseValue('bigint')).toBe(5608n);
 
@@ -425,6 +427,26 @@ describe('SwapKitNumber', () => {
     });
   });
 
+  describe('extending multiplier without loosing precision', () => {
+    test('edge case 1', () => {
+      const asset1 = new SwapKitNumber({ value: 41.90963702, decimal: 8 });
+      const multiplier = 5.337952274462478;
+      const divider = 105.2562773915526;
+      const result = asset1.mul(multiplier).div(divider);
+
+      expect(result.getValue('string')).toBe('2.12539953');
+    });
+
+    test('edge case 2', () => {
+      const asset1 = new SwapKitNumber('41.90963702');
+      const multiplier = new SwapKitNumber('5.337952274462478');
+      const divider = new SwapKitNumber('105.2562773915526');
+      const result = asset1.mul(multiplier).div(divider);
+
+      expect(result.getValue('string')).toBe('2.125399527674726');
+    });
+  });
+
   describe('gt', () => {
     test('greater than', () => {
       const skNumber1 = new SwapKitNumber(10);
@@ -432,6 +454,14 @@ describe('SwapKitNumber', () => {
 
       expect(skNumber1.gt(skNumber2)).toBe(true);
       expect(skNumber2.gt(skNumber1)).toBe(false);
+    });
+
+    test("different decimals doesn't affect comparison", () => {
+      const skNumber1 = new SwapKitNumber({ value: 10, decimal: 18 });
+      const skNumber2 = new SwapKitNumber({ value: '50', decimal: 8 });
+
+      expect(skNumber1.lt(skNumber2)).toBe(true);
+      expect(skNumber2.gt(skNumber1)).toBe(true);
     });
   });
 
@@ -475,6 +505,19 @@ describe('SwapKitNumber', () => {
       expect(skNumber1.eq(skNumber2)).toBe(false);
       expect(skNumber1.eq(skNumber1)).toBe(true);
       expect(skNumber2.eq(skNumber1)).toBe(false);
+    });
+  });
+
+  describe('comparison edge cases with decimals', () => {
+    test('compare on cut decimals', () => {
+      const skNumber1 = new SwapKitNumber({ value: 0.001, decimal: 3 });
+      const value = '0.0019';
+
+      expect(skNumber1.lt(value)).toBe(true);
+      expect(skNumber1.gt(value)).toBe(false);
+      expect(skNumber1.eq(value)).toBe(false);
+      expect(skNumber1.lte(value)).toBe(true);
+      expect(skNumber1.gte(value)).toBe(false);
     });
   });