+++ packages/swapkit/helpers/src/modules/assetValue.ts
@@ -21,12 +21,14 @@ import type { NumberPrimitives } from './bigIntArithmetics.ts';
 import { BigIntArithmetics, formatBigIntToSafeValue } from './bigIntArithmetics.ts';
 import type { SwapKitValueType } from './swapKitNumber.ts';
 
+type TokenTax = { buy: number; sell: number };
+
 const safeValue = (value: NumberPrimitives, decimal: number) =>
   typeof value === 'bigint'
     ? formatBigIntToSafeValue({ value, bigIntDecimal: decimal, decimal })
     : value;
 
-type AssetValueParams = { decimal: number; value: SwapKitValueType } & (
+type AssetValueParams = { decimal: number; value: SwapKitValueType; tax?: TokenTax } & (
   | { chain: Chain; symbol: string }
   | { identifier: string }
 );
@@ -46,7 +48,9 @@ type TokenNames =
   | (typeof WoofiList)['tokens'][number]['identifier']
   | (typeof UniswapList)['tokens'][number]['identifier'];
 
-let staticTokensMap: Map<TokenNames, { decimal: number; identifier: string }> | undefined;
+let staticTokensMap:
+  | Map<TokenNames, { tax?: TokenTax; decimal: number; identifier: string }>
+  | undefined;
 
 const getStaticToken = (identifier: TokenNames) => {
   if (!staticTokensMap) {
@@ -66,15 +70,19 @@ const createAssetValue = async (assetString: string, value: NumberPrimitives = 0
   return new AssetValue({ decimal, value: parsedValue, identifier: assetString });
 };
 
-const cacheAssetValue = new Map<string, AssetValue>();
-
 export class AssetValue extends BigIntArithmetics {
+  address?: string;
+  chain: Chain;
+  isGasAsset = false;
+  isSynthetic = false;
+  symbol: string;
+  tax?: TokenTax;
+  ticker: string;
+  type: ReturnType<typeof getAssetType>;
+
   constructor(params: AssetValueParams) {
     const identifier =
       'identifier' in params ? params.identifier : `${params.chain}.${params.symbol}`;
-
-    const cachedAsset = cacheAssetValue.get(identifier);
-    if (cachedAsset) return cachedAsset.set(params.value);
 
     super(
       params.value instanceof BigIntArithmetics
@@ -83,6 +91,7 @@ export class AssetValue extends BigIntArithmetics {
     );
 
     const assetInfo = getAssetInfo(identifier);
+
     this.type = getAssetType(assetInfo);
     this.chain = assetInfo.chain;
     this.ticker = assetInfo.ticker;
@@ -91,29 +100,21 @@ export class AssetValue extends BigIntArithmetics {
     this.isSynthetic = assetInfo.isSynthetic;
     this.isGasAsset = assetInfo.isGasAsset;
 
-    cacheAssetValue.set(identifier, this);
+    this.tax = params.tax;
   }
 
-  address?: string;
-  isSynthetic = false;
-  isGasAsset = false;
-  // @ts-expect-error cache is false positive on that case
-  chain: Chain;
-  // @ts-expect-error cache is false positive on that case
-  symbol: string;
-  // @ts-expect-error cache is false positive on that case
-  ticker: string;
-  // @ts-expect-error cache is false positive on that case
-  type: ReturnType<typeof getAssetType>;
-
   toString(short = false) {
-    // THOR.RUNE | ETH/ETH
-    const shortFormat = this.isSynthetic ? this.symbol.split('-')[0] : this.ticker;
+    const shortFormat = this.isSynthetic ? this.symbol : this.ticker;
 
     return short
-      ? shortFormat
+      ? // ETH/THOR-0xa5f2211b9b8170f694421f2046281775e8468044 | USDT
+        shortFormat
       : // THOR.ETH/ETH | ETH.USDT-0x1234567890
         `${this.chain}.${this.symbol}`;
+  }
+
+  toUrl() {
+    return this.isSynthetic ? `${this.chain}.${this.symbol.replace('/', '.')}` : this.toString();
   }
 
   eq({ chain, symbol }: { chain: Chain; symbol: string }) {
@@ -126,17 +127,19 @@ export class AssetValue extends BigIntArithmetics {
 
   static fromStringSync(assetString: string, value: NumberPrimitives = 0) {
     const { isSynthetic } = getAssetInfo(assetString);
-    const { decimal, identifier: tokenIdentifier } = getStaticToken(
-      assetString as unknown as TokenNames,
-    );
+    const {
+      tax,
+      decimal,
+      identifier: tokenIdentifier,
+    } = getStaticToken(assetString as unknown as TokenNames);
 
     const parsedValue = safeValue(value, decimal);
 
     const asset = tokenIdentifier
-      ? new AssetValue({ decimal, identifier: tokenIdentifier, value: parsedValue })
+      ? new AssetValue({ tax, decimal, identifier: tokenIdentifier, value: parsedValue })
       : isSynthetic
-      ? new AssetValue({ decimal: 8, identifier: assetString, value: parsedValue })
-      : undefined;
+        ? new AssetValue({ tax, decimal: 8, identifier: assetString, value: parsedValue })
+        : undefined;
 
     return asset;
   }
@@ -245,6 +248,6 @@ const getAssetInfo = (identifier: string) => {
     symbol:
       (isSynthetic ? `${synthChain}/` : '') +
       (address ? `${ticker}-${address?.toLowerCase() ?? ''}` : symbol),
-    ticker: ticker,
+    ticker,
   };
 };