+++ packages/toolboxes/cosmos/src/toolbox/gaia.ts
@@ -30,16 +30,34 @@ export const GaiaToolbox = ({ server }: { server?: string } = {}): GaiaToolboxTy
     client,
   });
 
+  const getFees = async () => {
+    const baseFee = (await getFeeRateFromThorswap(ChainId.Cosmos)) || 500;
+    return {
+      type: 'base',
+      average: SwapKitNumber.fromBigInt(BigInt(baseFee), BaseDecimal.GAIA),
+      fast: SwapKitNumber.fromBigInt((BigInt(baseFee) * 15n) / 10n, BaseDecimal.GAIA),
+      fastest: SwapKitNumber.fromBigInt(BigInt(baseFee) * 2n, BaseDecimal.GAIA),
+    };
+  };
+
   return {
     ...baseToolbox,
-    getFees: async () => {
-      const baseFee = (await getFeeRateFromThorswap(ChainId.Cosmos)) || 500;
-      return {
-        type: 'base',
-        average: new SwapKitNumber({ value: baseFee, decimal: BaseDecimal.GAIA }),
-        fast: new SwapKitNumber({ value: baseFee * 1.5, decimal: BaseDecimal.GAIA }),
-        fastest: new SwapKitNumber({ value: baseFee * 2, decimal: BaseDecimal.GAIA }),
-      };
+    getFees,
+    transfer: async (params: TransferParams) => {
+      const gasFees = await getFees();
+
+      return baseToolbox.transfer({
+        ...params,
+        fee: params.fee || {
+          amount: [
+            {
+              denom: 'uatom',
+              amount: gasFees[params.feeOptionKey || 'fast'].getBaseValue('string') || '1000',
+            },
+          ],
+          gas: '200000',
+        },
+      });
     },
   };
 };