+++ packages/toolboxes/utxo/src/index.ts
@@ -2,6 +2,7 @@ export { BaseUTXOToolbox } from './toolbox/BaseUTXOToolbox.ts';
 export { BCHToolbox } from './toolbox/bch.ts';
 export { BTCToolbox } from './toolbox/btc.ts';
 export { DOGEToolbox } from './toolbox/doge.ts';
+export { getToolboxByChain } from './toolbox/getToolboxByChain.ts';
 export { LTCToolbox } from './toolbox/ltc.ts';
 export * from './types/index.ts';
 export * from './utils/index.ts';