+++ packages/toolboxes/utxo/src/utils/coinselect.ts
@@ -32,7 +32,9 @@ export const accumulative = ({
 }: UTXOCalculateTxSizeParams & { outputs: TargetOutput[]; chain: UTXOChain }) => {
   feeRate = Math.ceil(feeRate);
   const newTxType =
-    'address' in inputs[0] ? getScriptTypeForAddress(inputs[0].address) : UTXOScriptType.P2PKH;
+    inputs[0] && 'address' in inputs[0]
+      ? getScriptTypeForAddress(inputs[0].address)
+      : UTXOScriptType.P2PKH;
   // skip input if adding it would cost more than input is worth
   const filteredInputs = inputs.filter((input) => getInputSize(input) * feeRate <= input.value);