+++ packages/wallets/wc/src/helpers.ts
@@ -2,10 +2,16 @@ import type { EVMChain } from '@swapkit/types';
 import { Chain } from '@swapkit/types';
 
 import {
+  ARBITRUM_ONE_MAINNET_ID,
   AVALACHE_MAINNET_ID,
   BINANCE_MAINNET_ID,
   BSC_MAINNET_ID,
+  COSMOS_HUB_MAINNET_ID,
   ETHEREUM_MAINNET_ID,
+  KUJIRA_MAINNET_ID,
+  MAYACHAIN_MAINNET_ID,
+  OPTIMISM_MAINNET_ID,
+  POLYGON_MAINNET_ID,
   THORCHAIN_MAINNET_ID,
 } from './constants.ts';
 
@@ -18,7 +24,7 @@ export const getAddressFromAccount = (account: string) => {
 };
 
 export const getAddressByChain = (
-  chain: EVMChain | Chain.Binance | Chain.THORChain,
+  chain: EVMChain | Chain.Binance | Chain.THORChain | Chain.Maya | Chain.Kujira | Chain.Cosmos,
   accounts: string[],
 ): string =>
   getAddressFromAccount(
@@ -37,6 +43,18 @@ export const chainToChainId = (chain: Chain) => {
       return BINANCE_MAINNET_ID;
     case Chain.THORChain:
       return THORCHAIN_MAINNET_ID;
+    case Chain.Arbitrum:
+      return ARBITRUM_ONE_MAINNET_ID;
+    case Chain.Optimism:
+      return OPTIMISM_MAINNET_ID;
+    case Chain.Polygon:
+      return POLYGON_MAINNET_ID;
+    case Chain.Maya:
+      return MAYACHAIN_MAINNET_ID;
+    case Chain.Cosmos:
+      return COSMOS_HUB_MAINNET_ID;
+    case Chain.Kujira:
+      return KUJIRA_MAINNET_ID;
     default:
       return '';
   }