+++ packages/wallets/xdefi/src/xdefiWallet.ts
@@ -1,3 +1,4 @@
+import type { AVAXToolbox, BSCToolbox } from '@swapkit/toolbox-evm';
 import {
   Chain,
   ChainId,
@@ -18,6 +19,7 @@ type XDEFIConfig = {
 };
 
 const XDEFI_SUPPORTED_CHAINS = [
+  Chain.Arbitrum,
   Chain.Avalanche,
   Chain.Binance,
   Chain.BinanceSmartChain,
@@ -27,6 +29,8 @@ const XDEFI_SUPPORTED_CHAINS = [
   Chain.Ethereum,
   Chain.Kujira,
   Chain.Litecoin,
+  Chain.Optimism,
+  Chain.Polygon,
   Chain.THORChain,
 ] as const;
 
@@ -74,31 +78,23 @@ const getWalletMethodsForChain = async ({
     case Chain.BitcoinCash:
     case Chain.Dogecoin:
     case Chain.Litecoin: {
-      const { BCHToolbox, BTCToolbox, DOGEToolbox, LTCToolbox } = await import(
-        '@swapkit/toolbox-utxo'
-      );
+      const { getToolboxByChain } = await import('@swapkit/toolbox-utxo');
       const params = { rpcUrl, blockchairApiKey, apiClient: api };
-      const toolbox =
-        chain === Chain.Bitcoin
-          ? BTCToolbox(params)
-          : chain === Chain.BitcoinCash
-          ? BCHToolbox(params)
-          : chain === Chain.Dogecoin
-          ? DOGEToolbox(params)
-          : LTCToolbox(params);
+      const toolbox = await getToolboxByChain(chain);
 
-      return { ...toolbox, transfer: walletTransfer };
+      return { ...toolbox(params), transfer: walletTransfer };
     }
 
     case Chain.Ethereum:
     case Chain.BinanceSmartChain:
+    case Chain.Arbitrum:
+    case Chain.Optimism:
+    case Chain.Polygon:
     case Chain.Avalanche: {
       const {
         getProvider,
         prepareNetworkSwitch,
-        ETHToolbox,
-        AVAXToolbox,
-        BSCToolbox,
+        getToolboxByChain,
         addEVMWalletNetwork,
         covalentApi,
         ethplorerApi,
@@ -117,20 +113,12 @@ const getWalletMethodsForChain = async ({
       }
 
       const provider = new BrowserProvider(ethereumWindowProvider, 'any');
-
-      const toolboxParams = {
+      const toolbox = (await getToolboxByChain(chain))({
         provider,
         signer: await provider.getSigner(),
         ethplorerApiKey: ethplorerApiKey || '',
         covalentApiKey: covalentApiKey || '',
-      };
-
-      const toolbox =
-        chain === Chain.Ethereum
-          ? ETHToolbox(toolboxParams)
-          : chain === Chain.Avalanche
-          ? AVAXToolbox(toolboxParams)
-          : BSCToolbox(toolboxParams);
+      });
 
       try {
         chain !== Chain.Ethereum &&