+++ internal/config/CHANGELOG.md
@@ -1,5 +1,57 @@
 # @internal/config
 
+## 0.0.0-rc.9
+
+### Patch Changes
+
+- Bump to sk links
+
+## 0.0.0-rc.8
+
+### Patch Changes
+
+- Update addLiquidityPart for symmetrics
+
+## 0.0.0-rc.7
+
+### Patch Changes
+
+- Bump for explorer & synths
+
+## 0.0.0-rc.6
+
+### Patch Changes
+
+- Bump
+
+## 0.0.0-rc.5
+
+### Patch Changes
+
+- tests
+
+- toUrl
+
+- [`9af9e284`](https://github.com/thorswap/SwapKit/commit/9af9e2845126818d4dace38457e219fffa1b3a8c) Thanks [@chillios-ts](https://github.com/chillios-ts)! - Update toString on synths
+
+## 0.0.0-rc.4
+
+### Patch Changes
+
+- Patch typeforce
+
+## 0.0.0-rc.3
+
+### Patch Changes
+
+- [#414](https://github.com/thorswap/SwapKit/pull/414) [`2c3f649f`](https://github.com/thorswap/SwapKit/commit/2c3f649fdebb5463e51c2929d6b3091852a59e9c) Thanks [@chillios-ts](https://github.com/chillios-ts)! - Fix Division precision
+
+## 0.0.0-rc.2
+
+### Patch Changes
+
+- Update dependencies with crypto-js 4.2.0
+
 ## 0.0.0-rc.1
 
 ### Patch Changes