+++ packages/swapkit/entities/CHANGELOG.md
@@ -1,5 +1,135 @@
 # @thorswap-lib/swapkit-entities
 
+## 1.0.0-rc.23
+
+### Patch Changes
+
+- Bump to sk links
+
+- Updated dependencies []:
+  - @swapkit/types@1.0.0-rc.23
+
+## 1.0.0-rc.22
+
+### Patch Changes
+
+- Update addLiquidityPart for symmetrics
+
+- Updated dependencies []:
+  - @swapkit/types@1.0.0-rc.22
+
+## 1.0.0-rc.21
+
+### Patch Changes
+
+- Bump for explorer & synths
+
+- Updated dependencies []:
+  - @swapkit/types@1.0.0-rc.21
+
+## 1.0.0-rc.20
+
+### Patch Changes
+
+- Bump
+
+- Updated dependencies []:
+  - @swapkit/types@1.0.0-rc.20
+
+## 1.0.0-rc.19
+
+### Patch Changes
+
+- tests
+
+- toUrl
+
+- [`9af9e284`](https://github.com/thorswap/SwapKit/commit/9af9e2845126818d4dace38457e219fffa1b3a8c) Thanks [@chillios-ts](https://github.com/chillios-ts)! - Update toString on synths
+
+- Updated dependencies [[`9af9e284`](https://github.com/thorswap/SwapKit/commit/9af9e2845126818d4dace38457e219fffa1b3a8c)]:
+  - @swapkit/types@1.0.0-rc.19
+
+## 1.0.0-rc.18
+
+### Patch Changes
+
+- Patch typeforce
+
+- Updated dependencies []:
+  - @swapkit/types@1.0.0-rc.18
+
+## 1.0.0-rc.17
+
+### Patch Changes
+
+- [#414](https://github.com/thorswap/SwapKit/pull/414) [`2c3f649f`](https://github.com/thorswap/SwapKit/commit/2c3f649fdebb5463e51c2929d6b3091852a59e9c) Thanks [@chillios-ts](https://github.com/chillios-ts)! - Fix Division precision
+
+- Updated dependencies [[`2c3f649f`](https://github.com/thorswap/SwapKit/commit/2c3f649fdebb5463e51c2929d6b3091852a59e9c)]:
+  - @swapkit/types@1.0.0-rc.17
+
+## 1.0.0-rc.16
+
+### Patch Changes
+
+- Update dependencies with crypto-js 4.2.0
+
+- Updated dependencies []:
+  - @swapkit/types@1.0.0-rc.16
+
+## 1.0.0-rc.15
+
+### Patch Changes
+
+- Testout publish
+
+- Updated dependencies []:
+  - @swapkit/types@1.0.0-rc.15
+
+## 1.0.0-rc.14
+
+### Patch Changes
+
+- Update package name generation
+
+- Updated dependencies []:
+  - @swapkit/types@1.0.0-rc.14
+
+## 1.0.0-rc.13
+
+### Patch Changes
+
+- [#404](https://github.com/thorswap/SwapKit/pull/404) [`af8ed16f`](https://github.com/thorswap/SwapKit/commit/af8ed16f77d15570f99a6062b7ba81273ff84b29) Thanks [@chillios-ts](https://github.com/chillios-ts)! - New chains support for evm/cosmos wallets
+
+- Updated dependencies [[`af8ed16f`](https://github.com/thorswap/SwapKit/commit/af8ed16f77d15570f99a6062b7ba81273ff84b29)]:
+  - @swapkit/types@1.0.0-rc.13
+
+## 1.0.0-rc.12
+
+### Patch Changes
+
+- bump packages
+
+- Updated dependencies []:
+  - @swapkit/types@1.0.0-rc.12
+
+## 1.0.0-rc.11
+
+### Patch Changes
+
+- addLiquidityPart and fix toCurrency
+
+- Updated dependencies []:
+  - @swapkit/types@1.0.0-rc.11
+
+## 1.0.0-rc.10
+
+### Patch Changes
+
+- remove cache asset
+
+- Updated dependencies []:
+  - @swapkit/types@1.0.0-rc.10
+
 ## 1.0.0-rc.9
 
 ### Patch Changes